feat: 모델 관계 설정 수정

+++ 1controllers/offer.controller.js
@@ -9,13 +9,16 @@ class OfferController {
 
     const { companyId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
     const { status, message } = await this.offerService.createOffer(
+
+      userId,
+      companyId,
       offerName,
-      offerOrder,
-      price,
-      companyId
+      offerNumber,
+      price
+
     );
     res.status(status).json(message);
   };
@@ -26,14 +29,17 @@ class OfferController {
 
     const { companyId, offerId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
     const { status, message } = await this.offerService.updateOffer(
       offerId,
+      companyId,
+      userId,
       offerName,
       offerOrder,
       price
     );
+
     res.status(status).json(message);
   };
 
@@ -41,9 +47,9 @@ class OfferController {
   deleteOffer = async (req, res, next) => {
     const { offerId } = req.params;
 
-    // const userId = res.locals.user
+    const userId = res.locals.userId;
 
-    const { status, message } = await this.offerService.destroyOffer(offerId);
+    const { status, message } = await this.offerService.destroyOffer(offerId, userId);
     res.status(status).json(message);
   };
